+++ examples/bsc-p2p/Cargo.toml
@@ -16,12 +16,18 @@ reth-network = { workspace = true, features = ["test-utils"] }
 reth-network-api.workspace = true
 reth-network-peers.workspace = true
 reth-primitives.workspace = true
+reth-primitives-traits.workspace = true
 reth-provider.workspace = true
 reth-tracing.workspace = true
 
 # alloy
 alloy-primitives.workspace = true
 alloy-rlp.workspace = true
+
+reth-engine-primitives.workspace = true
+reth-errors.workspace = true
+reth-payload-primitives.workspace = true
+alloy-rpc-types.workspace = true
 
 # misc
 bytes.workspace = true
